requirements.txt
notion-client==2.0.0
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.1.0
google-generativeai==0.3.2
kokoro>=0.8.4
soundfile>=0.12.1
numpy>=1.20.0
langchain==0.3.20